+++ src/main/kotlin/com/zepben/ewb/database/sqlite/cim/tables/extensions/iec61968/common/TableContactDetailsElectronicAddresses.kt
@@ -11,7 +11,7 @@ package com.zepben.ewb.database.sqlite.cim.tables.extensions.iec61968.common
 import com.zepben.ewb.database.sql.Column
 import com.zepben.ewb.database.sql.Column.Nullable.NOT_NULL
 import com.zepben.ewb.database.sql.Column.Type.STRING
-import com.zepben.ewb.database.sqlite.cim.tables.iec61968.common.TableElectronicAddress
+import com.zepben.ewb.database.sqlite.cim.tables.iec61968.common.TableElectronicAddresses
 
 /**
  * A class representing the ContactDetails to ElectronicAddress association columns required for the database table.
@@ -19,7 +19,7 @@ import com.zepben.ewb.database.sqlite.cim.tables.iec61968.common.TableElectronic
  * @property CONTACT_DETAILS_ID A column that stores the identifier of the contact details associated with the electronic address.
  */
 @Suppress("PropertyName")
-class TableContactDetailsElectronicAddresses : TableElectronicAddress() {
+class TableContactDetailsElectronicAddresses : TableElectronicAddresses() {
 
     val CONTACT_DETAILS_ID: Column = Column(++columnIndex, "contact_details_id", STRING, NOT_NULL)