+++ src/main/kotlin/com/zepben/ewb/database/sqlite/cim/tables/iec61968/common/TableAgreements.kt
@@ -15,13 +15,13 @@ import com.zepben.ewb.database.sql.Column.Type.STRING
 /**
  * A class representing the Agreement columns required for the database table.
  *
- * @property VALIDITY_INTERVAL_END A column storing the date and time interval this agreement is valid (from going into effect to termination).
- * @property VALIDITY_INTERVAL_START A column storing the date and time interval this agreement is valid (from going into effect to termination).
+ * @property VALIDITY_INTERVAL_START A column storing the date and time interval this agreement is valid (from going into effect...).
+ * @property VALIDITY_INTERVAL_END A column storing the date and time interval this agreement is valid (...to termination).
  */
 @Suppress("PropertyName")
 abstract class TableAgreements : TableDocuments() {
 
-    val VALIDITY_INTERVAL_END: Column = Column(++columnIndex, "validity_interval_end", STRING, NULL)
     val VALIDITY_INTERVAL_START: Column = Column(++columnIndex, "validity_interval_start", STRING, NULL)
+    val VALIDITY_INTERVAL_END: Column = Column(++columnIndex, "validity_interval_end", STRING, NULL)
 
 }